refactor(test): add support to use BASE_URL env value in all tests
