+++ src/tests/synchonizers/list-events.test.ts
@@ -1,10 +1,8 @@
-import { ListEventsResponse } from "../responses";
-import { Synchronizers } from "../index";
-import { EventInput, Options, Pagination, Sort } from "../types";
-import { baseAddress, baseEventInput, baseEventStatuses } from "./base";
+import { Synchronizers, Options, ListEventsResponse, EventInput } from "../../index";
+import { baseAddress, baseEventInput, baseEventStatuses } from "../base";
 
 describe("Synchronizers", () => {
-  const baseUrl = "http://localhost:5555";
+  const baseUrl = process.env.BASE_URL!;
   const eventInput: EventInput = baseEventInput;
   const addressInput = baseAddress;
   let synchronizers: Synchronizers;
